project/requirements.txt
Flask==3.0.3
gunicorn==20.1.0
Flask-CORS==5.0.0
Flask-Session==0.4.0 
python-dotenv>=0.19.2
authlib>=1.0
requests>=2.27.1